requirements.txt
pyenchant
pyyaml
praw